Sri/boring-todo/boring-todo-api/pyproject.toml
[project]
name = "boring-todo-api"
version = "0.1.0"
description = ""
authors = [{ name = "Sri Rang", email = "srirangan@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)", 
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "bcrypt (>=4.0.0,<5.0.0)",
    "email-validator (>=2.0.0,<3.0.0)"
]

[tool.poetry]
packages = [{ include = "boring_todo_api", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.3"
pytest-sugar = "^1.0.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v --tb=short --disable-warnings --cov=src/boring_todo_api --cov-report=term-missing --cov-report=xml"
